map et paiment ajoutés

+++ database/migrations/2021_02_24_233947_create_detail_products_table.php
@@ -17,8 +17,8 @@ class CreateDetailProductsTable extends Migration
         //$table->id();
             //$table->bigIncrements('id');
            
-            $table->unsignedBigInteger('productid');
-            $table->unsignedBigInteger('sellerid'); 
+            $table->Integer('productid');
+            $table->Integer('sellerid'); 
            // $table->Integer('stock');
             $table->foreign('sellerid')->references('sellerid')->on('sellers')->onDelete('cascade');
             $table->foreign('productid')->references('productid')->on('products')->onDelete('cascade');
